hopefully mitigated float errors (in theory)(probably)

+++ Group Project/interiorAngles.c
@@ -10,7 +10,7 @@
 
 
 
-double getAngle(float adjs1, float adjs2, float opps) {
+double getAngle(double adjs1, double adjs2, double opps) {
 	//cosine law
 	// C = cos^-1(  ( a^2 + b^2 - c^2 ) / (2ab)  )
 	//	  C	  cos^-1     a^2             b^2             -c^2                2ab
@@ -24,7 +24,7 @@ double getAngle(float adjs1, float adjs2, float opps) {
 }
 
 
-void calcAngles(float l1, float l2, float l3) {
+void calcAngles(double l1, double l2, double l3) {
 	//assumes lengths are already validated to form a triangle.
 
 	double a1 = getAngle(l2, l3, l1);
